backend/pyproject.toml

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.py$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
known_third_party = ["django", "rest_framework"]
skip = ["migrations"]


[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "profMsaif",email = "profmsaif@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "dj-database-url (>=3.0.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
]

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"
pre-commit = "^4.2.0"
flake8-import-order = "^0.19.1"
flake8-blind-except = "^0.2.1"
flake8-unused-arguments = "^0.0.13"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
